+++ src/arc/convert/__init__.py
@@ -1,6 +1,6 @@
 from arc.errors import ConversionError
-from arc.converter.base_converter import BaseConverter
-from arc.converter.converters import *
+from arc.convert.base_converter import BaseConverter
+from arc.convert.converters import *
 
 __all__ = [
     "StringConverter",
