+++ src/arc/script/__option.py
@@ -1,5 +1,5 @@
 from arc.config import Config
-from arc.converter.alias import convert_alias, is_alias
+from arc.convert.alias import convert_alias, is_alias
 from arc._utils import symbol
 
 NO_DEFAULT = symbol("NO_DEFAULT")
